Move helper methods into lemmatizer
