--- spacy/lang/ga/irish_morphology_helpers.py
@@ -1,94 +1,0 @@
-def demutate(word: str, is_hpref: bool = False) -> str:
-    UVOWELS = "AÁEÉIÍOÓUÚ"
-    LVOWELS = "aáeéiíoóuú"
-    lc = word.lower()
-    # remove eclipsis
-    if lc.startswith("bhf"):
-        word = word[2:]
-    elif lc.startswith("mb"):
-        word = word[1:]
-    elif lc.startswith("gc"):
-        word = word[1:]
-    elif lc.startswith("nd"):
-        word = word[1:]
-    elif lc.startswith("ng"):
-        word = word[1:]
-    elif lc.startswith("bp"):
-        word = word[1:]
-    elif lc.startswith("dt"):
-        word = word[1:]
-    elif word[0:1] == "n" and word[1:2] in UVOWELS:
-        word = word[1:]
-    elif lc.startswith("n-") and word[2:3] in LVOWELS:
-        word = word[2:]
-    # non-standard eclipsis
-    elif lc.startswith("bh-f"):
-        word = word[3:]
-    elif lc.startswith("m-b"):
-        word = word[2:]
-    elif lc.startswith("g-c"):
-        word = word[2:]
-    elif lc.startswith("n-d"):
-        word = word[2:]
-    elif lc.startswith("n-g"):
-        word = word[2:]
-    elif lc.startswith("b-p"):
-        word = word[2:]
-    elif lc.startswith("d-t"):
-        word = word[2:]
-
-    # t-prothesis
-    elif lc.startswith("ts"):
-        word = word[1:]
-    elif lc.startswith("t-s"):
-        word = word[2:]
-
-    # h-prothesis, if known to be present
-    elif is_hpref and word[0:1] == "h":
-        word = word[1:]
-    # h-prothesis, simple case
-    # words can also begin with 'h', but unlike eclipsis,
-    # a hyphen is not used, so that needs to be handled
-    # elsewhere
-    elif word[0:1] == "h" and word[1:2] in UVOWELS:
-        word = word[1:]
-
-    # lenition
-    # this breaks the previous if, to handle super-non-standard
-    # text where both eclipsis and lenition were used.
-    if lc[0:1] in "bcdfgmpst" and lc[1:2] == "h":
-        word = word[0:1] + word[2:]
-
-    return word
-
-
-def unponc(word: str) -> str:
-    # fmt: off
-    PONC = {
-        "ḃ": "bh",
-        "ċ": "ch",
-        "ḋ": "dh",
-        "ḟ": "fh",
-        "ġ": "gh",
-        "ṁ": "mh",
-        "ṗ": "ph",
-        "ṡ": "sh",
-        "ṫ": "th",
-        "Ḃ": "BH",
-        "Ċ": "CH",
-        "Ḋ": "DH",
-        "Ḟ": "FH",
-        "Ġ": "GH",
-        "Ṁ": "MH",
-        "Ṗ": "PH",
-        "Ṡ": "SH",
-        "Ṫ": "TH"
-    }
-    # fmt: on
-    buf = []
-    for ch in word:
-        if ch in PONC:
-            buf.append(PONC[ch])
-        else:
-            buf.append(ch)
-    return "".join(buf)
